Build fix.
svn:sync-xref-src-repo-rev: r113449

+++ src/VBox/Main/src-client/DrvAudioVRDE.cpp
@@ -1,4 +1,4 @@
-/* $Id: DrvAudioVRDE.cpp 65733 2017-02-10 15:37:38Z andreas.loeffler@oracle.com $ */
+/* $Id: DrvAudioVRDE.cpp 65734 2017-02-10 15:45:38Z andreas.loeffler@oracle.com $ */
 /** @file
  * VRDE audio backend for Main.
  */
@@ -507,7 +507,7 @@ static DECLCALLBACK(uint32_t) drvAudioVRDEStreamGetReadable(PPDMIHOSTAUDIO pInte
     {
         /* Return samples instead of bytes here
          * (since we specified PDMAUDIOSTREAMLAYOUT_RAW as the audio data layout). */
-        return PDMAUDIOSTREAMCFG_B2S(pStreamVRDE->pCfg, RTCircBufUsed(pStreamVRDE->In.pCircBuf));
+        return (uint32_t)PDMAUDIOSTREAMCFG_B2S(pStreamVRDE->pCfg, RTCircBufUsed(pStreamVRDE->In.pCircBuf));
     }
 
     return 0;
